+++ src/components/BodyContainer/index.js
@@ -10,11 +10,35 @@ class BodyContainer extends Component<void, Props, void> {
 
   render(): React$Element<any> {
     const { props }: { props: Props } = this
+    const { children, ...otherProps } = props
+
+    let child
+    if (typeof children === "string") {
+      child = children
+    }
+    else {
+      try {
+        child = React.Children.only(children)
+      }
+      catch (e) {
+        console.log("phenomic: BodyContainer: multiple childs")
+      }
+    }
+
+    if (child) {
+      return (
+        <div
+          className="phenomic-BodyContainer"
+          dangerouslySetInnerHTML={ { __html: child } }
+          { ...otherProps }
+        />
+      )
+    }
 
     return (
-      <div>
+      <div { ...otherProps }>
       {
-        React.Children.map(props.children, (child: any, i) => {
+        React.Children.map(children, (child: any, i) => {
           if (typeof child === "string") {
             return (
               <div